receiver fixes, receiver tests and some small fixes in kafka tests, database tests, and in database

+++ internal/services/receiver/handlers.go
@@ -32,6 +32,10 @@ func (r *Service) CreateProject(c echo.Context) error {
 		return c.JSON(http.StatusBadRequest, errMsg{Message: err.Error()})
 	}
 
+	if in.Name == "" || in.StartUrl == "" {
+		return c.JSON(http.StatusBadRequest, errMsg{Message: "invalid json"})
+	}
+
 	prj := models.Project{
 		Name:     in.Name,
 		StartUrl: in.StartUrl,
@@ -56,10 +60,18 @@ func (r *Service) CreateProject(c echo.Context) error {
 func (r *Service) GetProject(c echo.Context) error {
 	id := c.Param("id")
 
+	if id == "" {
+		return c.JSON(http.StatusBadRequest, errMsg{Message: "invalid id"})
+	}
+
 	prj, err := r.db.GetProject(id)
 
 	if errors.Is(err, models.DataBaseNotFound) {
 		return c.JSON(http.StatusNotFound, errMsg{Message: err.Error()})
+	}
+
+	if errors.Is(err, models.DataBaseWrongID) {
+		return c.JSON(http.StatusBadRequest, errMsg{Message: err.Error()})
 	}
 
 	if err != nil {
@@ -71,8 +83,13 @@ func (r *Service) GetProject(c echo.Context) error {
 }
 
 func (r *Service) GetAllShort(c echo.Context) error {
-	// prjs, err := r.db.GetShortProjectsByUserId()
-	panic("Implement GetShortProjectsByUserId in DataBase !!!")
+	prjs, err := r.db.GetProjectsByOwnerId(r.tempUUID)
+	if err != nil {
+		zap.S().Errorf("error while getting projects: %s", err)
+		return echo.ErrInternalServerError
+	}
+
+	return c.JSON(http.StatusOK, prjs)
 }
 
 type outDeleteProject struct {
@@ -81,6 +98,10 @@ type outDeleteProject struct {
 
 func (r *Service) DeleteProject(c echo.Context) error {
 	id := c.Param("id")
+
+	if id == "" {
+		return c.JSON(http.StatusBadRequest, errMsg{Message: "invalid id"})
+	}
 
 	err := r.db.DeleteProject(id)
